refactor(web): collapse the screen to one nav column + the room
The screen had three competing session-switching layers: the top bar
('N Plot Sessions'), the fleet rail (the actual list), and a per-session
'← all sessions' back link. The back link was redundant — the sidebar is
the switcher — and a symptom of chrome that repeats itself.

Collapse to two regions:

- The fleet rail becomes the single left chrome: brand (plot) + connection
  state (online/offline/handoff) folded into its header, then the fleet
  list. The top bar is deleted entirely — its content either moved here
  or was already redundant with the list.
- The room is just the session. The '← all sessions' back link is gone
  from the session surface and the snapshot-unavailable state; the rail
  owns switching. The only back affordance left is '← all work' / esc in
  focus mode — an intra-session mode exit, not session switching.

Landing on '/' now dives into the top session (the rail already shows the
fleet, so '/' is never a browse view); the OverviewPane is just the
empty/offline start state, no 'select a session' prompt.

RootLayout drops the top strip: the screen is flex(fleet | room) directly,
reclaiming the vertical layer. net -64 lines.
